requirements.txt
biopython
click
PyYAML
Rich
